cleanup

+++ projects/batfish-common-protocol/src/org/batfish/datamodel/routing_policy/expr/CommunitySetElem.java
@@ -42,7 +42,7 @@ public class CommunitySetElem implements Serializable {
          LiteralCommunitySetElemHalf suffix = (LiteralCommunitySetElemHalf) _suffix;
          int prefixInt = prefix.getValue();
          int suffixInt = suffix.getValue();
-         return (((long) prefixInt) << 16) | ((long) suffixInt);
+         return (((long) prefixInt) << 16) | (suffixInt);
       }
       else {
          throw new BatfishException("Does not represent a single community");
